Squash most GCC warnings

+++ examples/reduce/kd-tree.cpp
@@ -98,7 +98,7 @@ void print_block(void* b_, const diy::Master::ProxyWithLink& cp, void* verbose_)
                   link->bounds().max[0], link->bounds().max[1], link->bounds().max[2],
                   link->size(), b->points.size());
 
-  for (size_t i = 0; i < link->size(); ++i)
+  for (int i = 0; i < link->size(); ++i)
   {
       fprintf(stdout, "  (%d,%d,(%d,%d,%d)):",
                       link->target(i).gid, link->target(i).proc,
@@ -120,7 +120,7 @@ inline
 bool
 operator==(const diy::ContinuousBounds& x, const diy::ContinuousBounds& y)
 {
-    for (int i = 0; i < DIM; ++i)
+    for (unsigned i = 0; i < DIM; ++i)
     {
         if (x.min[i] != y.min[i])
             return false;
@@ -158,7 +158,7 @@ void verify_block(void* b_, const diy::Master::ProxyWithLink& cp, void* wrap_dom
   const Bounds& domain  = static_cast<WrapDomain*>(wrap_domain)->domain;
 
   for (size_t i = 0; i < b->points.size(); ++i)
-    for (int j = 0; j < DIM; ++j)
+    for (unsigned j = 0; j < DIM; ++j)
       if (b->points[i][j] < link->bounds().min[j] || b->points[i][j] > link->bounds().max[j])
         fprintf(stdout, "Warning: %f outside of [%f,%f] (%d)\n", b->points[i][j], link->bounds().min[j], link->bounds().max[j], j);
 
@@ -182,7 +182,7 @@ void verify_block(void* b_, const diy::Master::ProxyWithLink& cp, void* wrap_dom
   if (wrap)
       for (int i = 0; i < link->size(); ++i)
       {
-          for (int j = 0; j < DIM; ++j)
+          for (unsigned j = 0; j < DIM; ++j)
           {
             if (!(link->wrap(i)[j] == -1 && !(link->bounds().min[j] == domain.min[j] && link->bounds(i).max[j] == domain.max[j])) ||
                  (link->wrap(i)[j] ==  1 && !(link->bounds().max[j] == domain.max[j] && link->bounds(i).min[j] == domain.min[j])))
@@ -198,17 +198,17 @@ void verify_block(void* b_, const diy::Master::ProxyWithLink& cp, void* wrap_dom
 
   // verify that we intersect everybody in the link
   for (int i = 0; i < link->size(); ++i)
-      for (int j = 0; j < DIM; ++j)
+      for (unsigned j = 0; j < DIM; ++j)
       {
           if (!intersects(link->bounds(), link->bounds(i), j, wrap, domain))
               fprintf(stderr, "Warning: we don't intersect a block in the link: %d -> %d\n", cp.gid(), link->target(i).gid);
       }
 
   // verify that we don't intersect anybody not in the link
-  for (size_t i = 0; i < b->block_bounds.size(); ++i)
+  for (int i = 0; i < (int) b->block_bounds.size(); ++i)
   {
       if (i == cp.gid()) continue;
-      int j = 0;
+      unsigned j = 0;
       for (; j < DIM; ++j)
       {
           if (!intersects(link->bounds(), b->block_bounds[i], j, wrap, domain))
@@ -223,7 +223,7 @@ void verify_block(void* b_, const diy::Master::ProxyWithLink& cp, void* wrap_dom
                   break;
           }
           if (k == link->size())
-              fprintf(stderr, "Warning: we intersect a block not in the link: %d -/-> %lu\n", cp.gid(), i);
+              fprintf(stderr, "Warning: we intersect a block not in the link: %d -/-> %d\n", cp.gid(), i);
       }
   }
 }
