+++ include/diy/partners/common.hpp
@@ -74,7 +74,7 @@ fill_steps()
   {
     std::vector<int>    cur_steps(divisions().size(), 1);
 
-    for (int r = 0; r < rounds(); ++r)
+    for (size_t r = 0; r < rounds(); ++r)
     {
       steps_.push_back(cur_steps[kvs_[r].dim]);
       cur_steps[kvs_[r].dim] *= kvs_[r].size;
@@ -82,7 +82,7 @@ fill_steps()
   } else
   {
     std::vector<int>    cur_steps(divisions().begin(), divisions().end());
-    for (int r = 0; r < rounds(); ++r)
+    for (size_t r = 0; r < rounds(); ++r)
     {
       cur_steps[kvs_[r].dim] /= kvs_[r].size;
       steps_.push_back(cur_steps[kvs_[r].dim]);
@@ -153,7 +153,7 @@ factor(int k, const DivisionVector& divisions, KVSVector& kvs)
     bool changed = false;
     for (unsigned i = 0; i < divisions.size(); ++i)
     {
-      if (round_per_dim[i] == tmp_kvs[i].size())
+      if (round_per_dim[i] == (int) tmp_kvs[i].size())
         continue;
       kvs.push_back(DimK(i, tmp_kvs[i][round_per_dim[i]++]));
       changed = true;
